// packages/language-server/src/generated/ast.ts
export interface Attribute extends langium.AstNode {
    readonly $container: Entity | Relationship;
    readonly $type: 'Attribute';
    datatype?: DataType;
    name: string;
    type?: AttributeType;
    visibility?: VisibilityType;
}

export const Attribute = {
    $type: 'Attribute',
    datatype: 'datatype',
    name: 'name',
    type: 'type',
    visibility: 'visibility'
} as const;

export function isAttribute(item: unknown): item is Attribute {
    return reflection.isInstance(item, Attribute.$type);
}

export interface AttributeType extends langium.AstNode {
    readonly $container: Attribute;
    readonly $type: 'AttributeType';
    ATTR_NONE?: 'none';
    DERIVED?: 'derived';
    KEY?: 'key';
    MULTIVALUED?: 'multivalued';
    OPTIONAL?: 'optional';
    PARTIAL_KEY?: 'partial_key';
}

export const AttributeType = {
    $type: 'AttributeType',
    ATTR_NONE: 'ATTR_NONE',
    DERIVED: 'DERIVED',
    KEY: 'KEY',
    MULTIVALUED: 'MULTIVALUED',
    OPTIONAL: 'OPTIONAL',
    PARTIAL_KEY: 'PARTIAL_KEY'
} as const;

export function isAttributeType(item: unknown): item is AttributeType {
    return reflection.isInstance(item, AttributeType.$type);
}

export interface CardinalityType extends langium.AstNode {
    readonly $container: RelationEntity;
    readonly $type: 'CardinalityType';
    CARD_NONE?: 'none';
    MANY?: '1..N' | 'N';
    ONE?: '1' | '1..1';
    ZERO_OR_MORE?: '0..N';
    ZERO_OR_ONE?: '0..1';
}

export const CardinalityType = {
    $type: 'CardinalityType',
    CARD_NONE: 'CARD_NONE',
    MANY: 'MANY',
    ONE: 'ONE',
    ZERO_OR_MORE: 'ZERO_OR_MORE',
    ZERO_OR_ONE: 'ZERO_OR_ONE'
} as const;

export function isCardinalityType(item: unknown): item is CardinalityType {
    return reflection.isInstance(item, CardinalityType.$type);
}

export interface DataType extends langium.AstNode {
    readonly $container: Attribute;
    readonly $type: 'DataType';
    d?: number;
    size?: number;
    type: string;
}

export const DataType = {
    $type: 'DataType',
    d: 'd',
    size: 'size',
    type: 'type'
} as const;

export function isDataType(item: unknown): item is DataType {
    return reflection.isInstance(item, DataType.$type);
}

export interface Entity extends langium.AstNode {
    readonly $container: Model;
    readonly $type: 'Entity';
    attributes: Array<Attribute>;
    extends?: langium.Reference<Entity>;
    name: string;
    weak: boolean;
}

export const Entity = {
    $type: 'Entity',
    attributes: 'attributes',
    extends: 'extends',
    name: 'name',
    weak: 'weak'
} as const;

export function isEntity(item: unknown): item is Entity {
    return reflection.isInstance(item, Entity.$type);
}

export interface Model extends langium.AstNode {
    readonly $type: 'Model';
    entities: Array<Entity>;
    name: string;
    notation?: NotationOption;
    relationships: Array<Relationship>;
}

export const Model = {
    $type: 'Model',
    entities: 'entities',
    name: 'name',
    notation: 'notation',
    relationships: 'relationships'
} as const;

export function isModel(item: unknown): item is Model {
    return reflection.isInstance(item, Model.$type);
}

export interface NotationOption extends langium.AstNode {
    readonly $container: Model;
    readonly $type: 'NotationOption';
    notationType: NotationType;
}

export const NotationOption = {
    $type: 'NotationOption',
    notationType: 'notationType'
} as const;

export function isNotationOption(item: unknown): item is NotationOption {
    return reflection.isInstance(item, NotationOption.$type);
}

export interface NotationType extends langium.AstNode {
    readonly $container: NotationOption;
    readonly $type: 'NotationType';
    BACHMAN?: 'bachman';
    CHEN?: 'chen';
    CROWSFOOT?: 'crowfoot';
    NOTA_DEFAULT?: 'default';
    UML?: 'uml';
}

export const NotationType = {
    $type: 'NotationType',
    BACHMAN: 'BACHMAN',
    CHEN: 'CHEN',
    CROWSFOOT: 'CROWSFOOT',
    NOTA_DEFAULT: 'NOTA_DEFAULT',
    UML: 'UML'
} as const;

export function isNotationType(item: unknown): item is NotationType {
    return reflection.isInstance(item, NotationType.$type);
}

export interface RelationEntity extends langium.AstNode {
    readonly $container: RelationTarget | Relationship;
    readonly $type: 'RelationEntity';
    cardinality?: CardinalityType;
    entity: langium.Reference<Entity>;
    role?: string;
}

export const RelationEntity = {
    $type: 'RelationEntity',
    cardinality: 'cardinality',
    entity: 'entity',
    role: 'role'
} as const;

export function isRelationEntity(item: unknown): item is RelationEntity {
    return reflection.isInstance(item, RelationEntity.$type);
}

export interface Relationship extends langium.AstNode {
    readonly $container: Model;
    readonly $type: 'Relationship';
    attributes: Array<Attribute>;
    name: string;
    source: RelationEntity;
    targets: Array<RelationTarget>;
    weak: boolean;
}

export const Relationship = {
    $type: 'Relationship',
    attributes: 'attributes',
    name: 'name',
    source: 'source',
    targets: 'targets',
    weak: 'weak'
} as const;

export function isRelationship(item: unknown): item is Relationship {
    return reflection.isInstance(item, Relationship.$type);
}

export interface RelationshipType extends langium.AstNode {
    readonly $container: RelationTarget;
    readonly $type: 'RelationshipType';
    AGGREGATION_LEFT?: 'o-';
    AGGREGATION_RIGHT?: '-o';
    COMPOSITION_LEFT?: '*-';
    COMPOSITION_RIGHT?: '-*';
    RELA_DEFAULT?: '->';
}

export const RelationshipType = {
    $type: 'RelationshipType',
    AGGREGATION_LEFT: 'AGGREGATION_LEFT',
    AGGREGATION_RIGHT: 'AGGREGATION_RIGHT',
    COMPOSITION_LEFT: 'COMPOSITION_LEFT',
    COMPOSITION_RIGHT: 'COMPOSITION_RIGHT',
    RELA_DEFAULT: 'RELA_DEFAULT'
} as const;

export function isRelationshipType(item: unknown): item is RelationshipType {
    return reflection.isInstance(item, RelationshipType.$type);
}

export interface RelationTarget extends langium.AstNode {
    readonly $container: Relationship;
    readonly $type: 'RelationTarget';
    relationEntity: RelationEntity;
    type: RelationshipType;
}

export const RelationTarget = {
    $type: 'RelationTarget',
    relationEntity: 'relationEntity',
    type: 'type'
} as const;

export function isRelationTarget(item: unknown): item is RelationTarget {
    return reflection.isInstance(item, RelationTarget.$type);
}

export interface VisibilityType extends langium.AstNode {
    readonly $container: Attribute;
    readonly $type: 'VisibilityType';
    PACKAGE?: 'package' | '~';
    PRIVATE?: '-' | 'private';
    PROTECTED?: '#' | 'protected';
    PUBLIC?: '+' | 'public';
    VISI_NONE?: 'none';
}

export const VisibilityType = {
    $type: 'VisibilityType',
    PACKAGE: 'PACKAGE',
    PRIVATE: 'PRIVATE',
    PROTECTED: 'PROTECTED',
    PUBLIC: 'PUBLIC',
    VISI_NONE: 'VISI_NONE'
} as const;